backend/pyproject.toml
[project]
name = "scripe"
version = "0.1.0"
description = "B2B Lead Generation Tool - Compliance-First Approach"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # HTTP & Scraping
    "httpx>=0.27.0",
    "selectolax>=0.3.21",
    "beautifulsoup4>=4.12.3",
    # Database
    "sqlalchemy>=2.0.25",
    # Data validation
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    # CLI
    "typer>=0.12.3",
    "rich>=13.7.0",
    # Logging
    "structlog>=24.1.0",
    # Phone validation
    "phonenumbers>=8.13.27",
    # Utils
    "python-slugify>=8.0.1",
    "tenacity>=8.2.3",
    # API
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    # Export
    "openpyxl>=3.1.2",
    "reportlab>=4.0.8",
    # Auth
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "dnspython>=2.4.2",  # For email MX validation
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.3",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.14",
    "mypy>=1.8.0",
]

[project.scripts]
scripe = "scripe.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
